move fetchBalance to utils folder
